refactor: remove components/controller workpsace
