--- components/controller/src/lib.rs
@@ -1,5 +1,0 @@
-// pub mod wallet_feature;
-pub mod controller;
-
-#[cfg(test)]
-mod test;
