+++ components/witness/Cargo.toml
@@ -13,9 +13,8 @@ figment = {version = "0.10.6", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 url = { version = "2.2.2", features = ["serde"] }
 reqwest = { version = "0.11", features = ["json", "rustls-tls"], default-features = false }
-keri = {path = "../../keriox_core", features = ["oobi"]}
+keri = {path = "../../keriox_core", features = ["oobi", "controller"]}
 tempfile = { version = "3.1"}
 
 [dev-dependencies]
-controller = {path = "../controller"}
 serde_json = "1.0"
